+++ hypothesis-python/tests/nocover/test_argument_validation.py
@@ -18,7 +18,7 @@ from inspect import Parameter
 import pytest
 
 from hypothesis import strategies as st
-from hypothesis.strategies._internal.core import _strategies
+from hypothesis.strategies._internal.utils import _strategies
 
 from tests.common.arguments import argument_validation_test, e
